refactor!: separate mineflayer events into folders
