+++ src/events/mineflayer/chat/memberKicked.ts
@@ -1,6 +1,6 @@
 import { ChatMessage } from "prismarine-chat";
-import Emojis from "../../util/emojis";
-import { Event } from "../../interfaces/Event";
+import Emojis from "../../../util/emojis";
+import { Event } from "../../../interfaces/Event";
 import { Util } from "discord.js";
 
 export default {
