+++ src/events/mineflayer/chat/questComplete.ts
@@ -1,5 +1,5 @@
-import Emojis from "../../util/emojis";
-import { Event } from "../../interfaces/Event";
+import Emojis from "../../../util/emojis";
+import { Event } from "../../../interfaces/Event";
 import { MessageEmbed } from "discord.js";
 
 export default {
