+++ src/events/mineflayer/chat/sameMessageTwice.ts
@@ -1,4 +1,4 @@
-import { Event } from "../../interfaces/Event";
+import { Event } from "../../../interfaces/Event";
 import { MessageEmbed } from "discord.js";
 
 export default {
