+++ src/events/mineflayer/handler/error.ts
@@ -1,4 +1,4 @@
-import { Event } from "../../interfaces/Event";
+import { Event } from "../../../interfaces/Event";
 
 export default {
 	name: "error",
